Frustum culling should be working.
