+++ include/Engine/Rendering/RenderSystem.h
@@ -31,9 +31,6 @@ private:
     const IRenderer* m_Renderer;
     RenderFrame* m_RenderFrame;
     Camera* m_Camera;
-    Camera* m_LastCullCamera;
-    Camera** m_FrustumCamPtr;
-    EntityWrapper frustumEntity;
     World* m_World;
     EntityWrapper m_CurrentCamera = EntityWrapper::Invalid;
     EntityWrapper m_LocalPlayer = EntityWrapper::Invalid;
